requirements.txt
pinecone[grpc]>=7.3.0
sentence-transformers>=2.3.0
pandas>=2.1.0
numpy>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0